requirements.txt
streamlit>=1.38.0
python-dotenv>=1.0.0
azure-identity>=1.19.0
azure-ai-projects>=1.0.0b5
openai>=1.55.0
